add multi dataset prepare
